+++ sessions.php
@@ -33,6 +33,7 @@ require_once __DIR__ . '/includes/system/class-environment.php';
 require_once __DIR__ . '/autoload.php';
 require_once __DIR__ . '/includes/libraries/class-libraries.php';
 require_once __DIR__ . '/includes/libraries/autoload.php';
+require_once __DIR__ . '/includes/features/class-wpcli.php';
 
 /**
  * The code that runs during plugin activation.